requirements.txt
ipython==8.12.3
matplotlib==3.5.3
numpy==1.23.1
pandas==2.2.3
plotly==5.24.1
protobuf==5.28.3
python-dotenv==1.0.1
Requests==2.32.3
scipy==1.9.0
seaborn==0.13.2
streamlit==1.39.0
yfinance==0.2.48
